config/bigquery/requirements.txt
pandas
sqlalchemy
python-dotenv
google-cloud-bigquery
mysql-connector-python
pyarrow
pandas-gbq>=0.26.1